refactor(ext): move extension lifecycle handling into ExtensionService
Rename ExtensionSupervisor to ExtensionService and move the extension-facing lifecycle handling out of ExtensionMode into it: the startup of enabled extensions, the batch start/stop jobs, and the eventbus handlers for the D-Bus intents (extensions:reload, extensions:stop, extensions:preview_ext, extensions:stop_preview).

ExtensionMode had no business handling the extension lifecycle. After this split it is a pure Mode: it handles queries, triggers and the active extension conversation, and only activates the service on init.

Both classes keep their own EventBus("extensions") instance. Listeners are registered per full event name and each bus carries its own bound instance, so the D-Bus wire names stay unchanged. The service bus uses skip_if_not_bound so unbound lifecycle events are dropped in processes that never activate the service (the CLI).

No behavior change.
