+++ tests/modes/extensions/test_extension_registry.py
@@ -13,7 +13,7 @@ def test_iterate__orders_preview_enabled_error_disabled(mocker: MockerFixture) -
     disabled = SimpleNamespace(id="disabled", is_preview=False, is_enabled=False, has_error=False)
 
     controllers = {c.id: c for c in (disabled, errored, enabled)}
-    mocker.patch.object(extension_registry.supervisor, "_preview", preview)
+    mocker.patch.object(extension_registry.ext_service, "_preview", preview)
     mocker.patch.object(
         extension_registry.extension_finder,
         "iterate",
